LPS-53821 Rename package [no logic changes]
