+++ modules/apps/item-selector/item-selector-criteria-api/bnd.bnd
@@ -1,5 +1,5 @@
-Bundle-Name: Liferay Document Library Item Selector API
-Bundle-SymbolicName: com.document.library.item.selector.api
+Bundle-Name: Liferay Item Selector Criteria API
+Bundle-SymbolicName: com.liferay.item.selector.criteria.api
 Bundle-Version: 1.0.0
-Export-Package: com.liferay.document.library.item.selector
+Export-Package: com.liferay.item.selector.criteria
 Include-Resource: classes
