+++ modules/apps/item-selector/item-selector-criteria-api/src/com/liferay/item/selector/criteria/DLItemSelectorCriterion.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.document.library.item.selector;
+package com.liferay.item.selector.criteria;
 
 import com.liferay.item.selector.BaseItemSelectorCriterion;
 import com.liferay.portal.kernel.repository.model.FileEntry;
